refactor(scalars): move tooltip to new UI folder
