+++ packages/design-system/src/scalars/components/fragments/id-autocomplete-field/id-autocomplete-input-container.tsx
@@ -4,8 +4,11 @@ import { cn } from "#scalars";
 import { Command as CommandPrimitive } from "cmdk";
 import React, { useState } from "react";
 import { useMediaQuery } from "usehooks-ts";
+import {
+  Tooltip,
+  TooltipProvider,
+} from "../../../../ui/components/tooltip/tooltip.js";
 import { Input } from "../input/input.js";
-import { Tooltip, TooltipProvider } from "../tooltip/tooltip.js";
 import type { IdAutocompleteOption } from "./types.js";
 
 interface IdAutocompleteInputContainerProps {
