+++ packages/design-system/src/ui/components/index.ts
@@ -21,3 +21,6 @@ export {
   type SidebarNode,
   type SidebarProps,
 } from "./sidebar/index.js";
+
+// TODO: export tooltip once it is ready to be used outside the package
+// DO NOT export tooltip until it is ready to be used outside the package
