+++ packages/design-system/src/ui/components/tooltip/tooltip.tsx
@@ -10,7 +10,7 @@ import {
   Trigger,
 } from "@radix-ui/react-tooltip";
 import * as React from "react";
-import { cn } from "../../../lib/utils.js";
+import { cn } from "../../../scalars/lib/utils.js";
 
 interface TooltipProps
   extends TooltipPrimitiveProps,
